refactor: move form to widget
