+++ index.html
@@ -7,7 +7,7 @@
     <meta name="description" content="React Forms" />
     <title>React Forms</title>
   </head>
-  <body>
+  <body class="scrollbar">
     <div id="root"></div>
     <script type="module" src="src"></script>
   </body>
