+++ src/app/redux/store.ts
@@ -1,10 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit'
 import { countriesReducer } from './slices/countries.slice'
-import { firstReducer } from './slices/first.slice'
+import { submitHistoryReducer } from './slices/first.slice'
 
 export const store = configureStore({
   reducer: {
-    firstForm: firstReducer,
+    submitHistory: submitHistoryReducer,
     countries: countriesReducer,
   },
 })
