+++ src/globals.css
@@ -67,11 +67,6 @@ input[type='email'],
 input[type='password'] {
   padding: 0.25rem 0.4rem;
   border: 1px solid #000;
-
-  &:invalid,
-  &.invalid {
-    border: 1px solid #f00;
-  }
 }
 
 input[type='checkbox'] {
@@ -108,4 +103,46 @@ input[type='checkbox'] {
 
   font-size: 0.66rem;
   color: #f00;
+}
+
+.tile-list {
+  display: flex;
+  flex-direction: column-reverse;
+  width: fit-content;
+  list-style: none;
+}
+
+.tile {
+  width: 100%;
+  padding: 1rem;
+  box-shadow: inset 0 0 0.1rem 0.1rem teal;
+}
+
+.tile-list > .tile-list-item:last-child > .tile {
+  animation: 1s teal-glow linear 3;
+}
+
+.tile-title {
+  font-weight: 500;
+  text-decoration: underline;
+}
+
+@keyframes teal-glow {
+  0% {
+    box-shadow:
+      inset 0 0 0.1rem 0.1rem teal,
+      0 0 0.25rem 0.25rem #00808000;
+  }
+
+  50% {
+    box-shadow:
+      inset 0 0 0.1rem 0.1rem teal,
+      0 0 0.25rem 0.25rem #008080ff;
+  }
+
+  100% {
+    box-shadow:
+      inset 0 0 0.1rem 0.1rem teal,
+      0 0 0.25rem 0.25rem #00808000;
+  }
 }
